Handle SIGINT when running native code.
Python has a fairly poor handling of SIGINT (Ctrl+C) in multi-threaded
code.
The way we deal with it is to replace the python SIGINT handler with the
OS default one (that kills the process) when running long computations
implemented in native code.
The main drawback is that we get an impproper shutdown, but that should
not be too much of an issue.

Fixes #12