Do not build BLIS by default on linux.
BLIS provides only a modest improvement over matrixmultiply, and is a
rather annoying dependency to deal with (slow to compile, works only on
some platforms...).
Moreover, SCALib is (or shouldn't) generally be limited by
floating-point matrix multiplication performance, since we mostly work
with i16 data.

This is a first step towards removal.