Scaled histogram for better accuracy in rank estimation (#184)
This new method avoids catastrophic cancellations in the FFT computations by scaling histogram bin counts by an appropriate amount. Each histogram is scaled by a factor, which is properly tracked. Further, we both floor and ceil when scaling bin counts, giving two histograms as lower- and upper-bound of the true histogram.

This scaling and rounding should not lead to significant loss in relative accuracy in the final rank (it is nevertheless properly reported in the final bounds).

---------

Co-authored-by: Gaëtan Cassiers <gaetan.cassiers@uclouvain.be>